refactor(Exchange): prefix shapeshift files with 'Partner'
