+++ Blockchain/Blockchain-Bridging-Header.h
@@ -52,7 +52,7 @@
 
 #import "FromToView.h"
 #import "ContinueButtonInputAccessoryView.h"
-#import "ExchangeOverviewViewController.h"
+#import "PartnerExchangeListViewController.h"
 
 #import "ExchangeTrade.h"
 #import "NSDateFormatter+TimeAgoString.h"
