+++ Blockchain/View Controllers/PartnerExchangeCreateViewController.m
@@ -6,7 +6,7 @@
 //  Copyright © 2017 Blockchain Luxembourg S.A. All rights reserved.
 //
 
-#import "ExchangeCreateViewController.h"
+#import "PartnerExchangeCreateViewController.h"
 #import "FromToView.h"
 #import "Blockchain-Swift.h"
 #import "ContinueButtonInputAccessoryView.h"
@@ -20,7 +20,7 @@
 
 #define IMAGE_NAME_SWITCH_CURRENCIES @"switch_currencies"
 
-@interface ExchangeCreateViewController () <UITextFieldDelegate, AddressSelectionDelegate, ContinueButtonInputAccessoryViewDelegate, ExchangeCreateViewDelegate>
+@interface PartnerExchangeCreateViewController () <UITextFieldDelegate, AddressSelectionDelegate, ContinueButtonInputAccessoryViewDelegate, ExchangeCreateViewDelegate>
 
 @property (nonatomic) NSTimer *quoteTimer;
 
@@ -51,7 +51,7 @@
 @property (nonatomic) FromToButtonDelegateIntermediate *fromToButtonDelegateIntermediate; // strong reference required to prevent it from becoming nil
 @end
 
-@implementation ExchangeCreateViewController
+@implementation PartnerExchangeCreateViewController
 
 - (void)viewDidLoad
 {
